Refactored the create subcommand
- Also organized the CLI tests
